discord/requirements.txt
discord
py-cord==2.6
python-dotenv
unidecode